envoy: Use string name to identify policy.
Different endpoints sharting the same security identity can have a
different policy due to endpoint specific enforcement modes. Each
proxy redirect is fundamentally bount to a specific cilium endpoint
(pod), so we need to use stable reference to the endpoint to configure
the policy for the proxy redirect. This commit uses a string
representation of the Cilium Endpoint ID for this purpose.

Envoy API protocols have a convention of identifying API resouces with
a string name. Use the same endpoint policy name for this purpose as
well.

Signed-off-by: Jarno Rajahalme <jarno@covalent.io>